Move Network resource to internal and begin development of search data source
